python-app/requirements.txt
pika
google-cloud-storage
